remove error handling implementations
After some conversations with colleagues, the error handling
implementation provided here is too restrictive. The magic lies in the
different types and wrapping functions. There are too many different
ways a project may need to handle the errors returned from handlers.

A project could:
- have different encodings
- different localizations

just to name a few that came up.

A project importing this (or copy pasting it to avoid a little
dependency) will need to implement it's own way of handling errors.